Make manim into a module and refactor configuration
